% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_utils.R
\name{convert_modkit_bed_to_parquet_hive}
\alias{convert_modkit_bed_to_parquet_hive}
\title{Fahrenheit conversion2}
\usage{
convert_modkit_bed_to_parquet_hive(file_path, output_dir)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
